由来源插件提供接纳控制和待回复判断
