+++ plugin_boundary_baseline.toml
@@ -54,10 +54,8 @@ R2 = [
     "plugins/codex/responses.py|core.net.http",
     "plugins/content/api.py|agent.turn_effects",
     "plugins/content/api.py|session.artifacts",
-    "plugins/conversation/plugin.py|agent.restart",
     "plugins/conversation/program.py|agent.model_runtime.session_selection",
     "plugins/conversation/source.py|agent.model_runtime.session_selection",
-    "plugins/conversation/source.py|agent.restart",
     "plugins/delivery_policy/plugin.py|agent.restart",
     "plugins/markdown_memory/message_plugin.py|agent.llm_json",
     "plugins/markdown_memory/message_plugin.py|agent.turn_effects",
@@ -77,7 +75,6 @@ R2 = [
     "plugins/programmatic/control.py|agent.control.frame_book",
     "plugins/programmatic/control.py|agent.control.protocol.models",
     "plugins/programmatic/plugin.py|agent.control.frame_book",
-    "plugins/programmatic/plugin.py|agent.restart",
     "plugins/prompt/plugin.py|agent.persona",
     "plugins/prompt/text.py|agent.persona",
     "plugins/prompt/text.py|prompts.completion",
@@ -172,7 +169,6 @@ R3 = [
     "plugins/conversation/plugin.py|plugins.content.api",
     "plugins/conversation/plugin.py|plugins.content.plugin",
     "plugins/conversation/plugin.py|plugins.models.selection",
-    "plugins/conversation/plugin.py|plugins.sources.plugin",
     "plugins/conversation/program.py|plugins.content.plugin",
     "plugins/conversation/program.py|plugins.context.api",
     "plugins/conversation/program.py|plugins.context.materials",
@@ -236,9 +232,7 @@ R3 = [
     "plugins/programmatic/control.py|plugins.turn_projection.plugin",
     "plugins/programmatic/plugin.py|plugins.content.plugin",
     "plugins/programmatic/plugin.py|plugins.conversation.plugin",
-    "plugins/programmatic/plugin.py|plugins.conversation.source",
     "plugins/programmatic/plugin.py|plugins.delivery.api",
-    "plugins/programmatic/plugin.py|plugins.sources.plugin",
     "plugins/programmatic/plugin.py|plugins.turn_projection.plugin",
     "plugins/programmatic/result.py|plugins.turn_projection.plugin",
     "plugins/react/plugin.py|plugins.content.plugin",
@@ -247,17 +241,14 @@ R3 = [
     "plugins/react/plugin.py|plugins.models.projection",
     "plugins/react/plugin.py|plugins.tools.menu",
     "plugins/reply/api.py|plugins.context.api",
-    "plugins/reply/follow.py|plugins.sources.plugin",
     "plugins/reply/plugin.py|plugins.content.plugin",
     "plugins/reply/plugin.py|plugins.context.api",
     "plugins/reply/plugin.py|plugins.context.materials",
     "plugins/reply/plugin.py|plugins.context.plugin",
     "plugins/reply/plugin.py|plugins.conversation.commands",
     "plugins/reply/plugin.py|plugins.conversation.program",
-    "plugins/reply/plugin.py|plugins.conversation.source",
     "plugins/reply/plugin.py|plugins.models.projection",
     "plugins/reply/plugin.py|plugins.react.plugin",
-    "plugins/reply/plugin.py|plugins.sources.plugin",
     "plugins/reply/plugin.py|plugins.tool_search.plugin",
     "plugins/reply/plugin.py|plugins.tools.api",
     "plugins/reply/plugin.py|plugins.tools.plugin",
