+++ plugins/reply/follow.py
@@ -8,11 +8,26 @@ from dataclasses import dataclass
 from agent.plugin_composition import Context
 from agent.plugin_composition.tasks import Task
 from agent.restart import RestartGate
-from plugins.sources.plugin import Source, Sources
+from typing import Protocol
 from agent.plugin_composition.messages import MessageCatalog, MessageReader
 
 logger = logging.getLogger(__name__)
 Program = Callable[[Task, MessageReader, str], Awaitable[object]]
+
+
+class SourceSession(Protocol):
+    async def start(self, program: Program) -> Task | None: ...
+
+
+class Source(Protocol):
+    @property
+    def name(self) -> str: ...
+    def open(self, session_id: str) -> SourceSession: ...
+
+
+class Sources(Protocol):
+    def entries(self) -> tuple[Source, ...]: ...
+    def needs_reply(self, reader: MessageReader, source: str) -> bool: ...
 
 
 @dataclass(slots=True)
