+++ plugins/sources/plugin.py
@@ -4,9 +4,8 @@ from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
 from typing import Protocol
 
-from agent.plugin_composition import Context, ServiceKey
+from agent.plugin_composition import Context, Effect, ServiceKey
 from agent.plugin_composition.channels import CHANNEL_INPUT, ChannelInboundMessage
-from agent.plugin_composition.effect import Effect
 from agent.plugin_composition.tasks import Task
 from agent.plugin_composition.messages import MessageReader
 from agent.plugin_contracts import Message
@@ -21,11 +20,12 @@ Accept = Callable[[str, str, ChannelInboundMessage], Awaitable[Message]]
 Changed = Callable[[MessageReader, str], None]
 
 
-class SourceSession(Protocol):
-    """来源注册表只需要启动一个已打开 Session 的程序。"""
+from .session import SourceSession as _SourceSession
 
+
+class SourceSession(Protocol):
     async def start(
-        self, program: Callable[[Task, MessageReader, str], Awaitable[object]]
+        self, program: Callable[[Task, MessageReader, str], Awaitable[object]],
     ) -> Task | None: ...
 
 
@@ -33,6 +33,7 @@ class SourceSession(Protocol):
 class Source:
     name: str
     open: Callable[[str], SourceSession]
+    needs_reply: Callable[[MessageReader], bool]
     accept: Accept | None = None
     channels: tuple[str, ...] | None = ()
 
@@ -43,8 +44,13 @@ class Sources:
     def __init__(self):
         self._items: dict[str, Source] = {}
 
-    async def register(self, ctx: Context, source: Source) -> Effect:
+    async def register(
+        self, ctx: Context, *, name: str, open: Callable[[str], SourceSession],
+        needs_reply: Callable[[MessageReader], bool], accept: Accept | None = None,
+        channels: tuple[str, ...] | None = (),
+    ) -> Effect:
         """来源注册随插件 effect 生灭；None channels 表示唯一默认输入来源。"""
+        source = Source(name, open, needs_reply, accept, None if channels is None else tuple(channels))
         if ctx.require(SOURCES) is not self:
             raise PermissionError("来源注册不属于当前组合")
         if not source.name or source.accept is None and source.channels != () or (
@@ -70,6 +76,10 @@ class Sources:
 
         return await ctx.effect(setup, label="source:" + source.name)
 
+    def needs_reply(self, reader: MessageReader, source: str) -> bool:
+        """来源自己解释是否待回复；注册表不猜来源的完成规则。"""
+        return self._items[source].needs_reply(reader)
+
     def entries(self) -> tuple[Source, ...]:
         return tuple(self._items.values())
 
@@ -88,11 +98,13 @@ class Sources:
         return await default.accept(session_id, message_id, message)
 
 
-SOURCES = ServiceKey[Sources]("sources.v1")
+SOURCES = ServiceKey[Sources]("sources.v2")
+SOURCE_SESSION = ServiceKey[type[_SourceSession]]("source.session.v1")
 SOURCE_CHANGED = ServiceKey[Changed]("source.changed.v1")
 
 
 async def apply(ctx: Context, config: object) -> None:
     sources = Sources()
+    _ = await ctx.provide(SOURCE_SESSION, _SourceSession)
     _ = await ctx.provide(SOURCES, sources)
     _ = await ctx.provide(CHANNEL_INPUT, sources.accept)
